Have Font Resolution Multiplier be a variable instead of hard coded
Totally understand if this is decided to be not incorporated, but just wanted to raise the point now that I have chatted with a few more FotG folks. Currently our base font size, and thus what the players see as their first impression, is a at 16 size for 1080. I know the 0.012 value was based on standard practice for font resolution methods used, though I just wanted to check if you might be open to just moving the 0.012 to a variable here, that way mods can set the base font resolution multiplier exactly to the kinds of fonts they might be using. Again happy to withdraw PR if this is something you don't want in the main repo!